refactor(steps-4-5): extract PlaybackQueue and TrackGenerator
Step 4: PlaybackQueue class in voxd/playback.py (417 lines) with
_play_audio, _probe_duration, _player_command, consumer loop. Moves
playback_queue and last_playback state off DaemonContext with delegation
properties. 6 test classes moved to test_voxd_playback.py.

Step 5: TrackGenerator class in voxd/track_generator.py (95 lines)
with generate, auto_track_name, list_tracks, slugify. 13 tests in
test_voxd_track_gen.py.

_monolith.py: 1922→1645 lines (-277). method_ratio +0.070,
encapsulation_ratio +0.056, avg_complexity -0.170,
classes_per_module 3→2, public_attr_violations -2.
Test count: 1453→1466 (+13).
