+++ .oo-audit.jsonl
@@ -15,3 +15,8 @@
 {"ts":"2026-05-14T00:33:51Z","commit":"6223355","files_scored":39,"files_improved":3,"files_regressed":1,"verdict":"fail","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[102.0,106.0]},"src/punt_vox/voxd/chimes.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.5],"max_complexity":[0.0,4.0],"avg_complexity":[0.0,2.5],"module_size":[0.0,42.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/dedup.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.0],"max_complexity":[0.0,8.0],"avg_complexity":[0.0,3.2],"module_size":[0.0,165.0],"classes_per_module":[0.0,3.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,2.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
 {"ts":"2026-05-14T00:35:08Z","commit":"6223355","files_scored":39,"files_improved":4,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[102.0,106.0]},"src/punt_vox/voxd/_monolith.py":{"avg_params":[1.9,2.23],"avg_complexity":[3.71,3.88],"module_size":[2330.0,1922.0],"classes_per_module":[7.0,3.0],"init_violations":[3.0,1.0]},"src/punt_vox/voxd/chimes.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.5],"max_complexity":[0.0,4.0],"avg_complexity":[0.0,2.5],"module_size":[0.0,42.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/dedup.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.0],"max_complexity":[0.0,8.0],"avg_complexity":[0.0,3.2],"module_size":[0.0,165.0],"classes_per_module":[0.0,3.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,2.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
 {"ts":"2026-05-14T00:54:24Z","commit":"11fd6a3","files_scored":39,"files_improved":3,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[106.0,110.0]},"src/punt_vox/voxd/chimes.py":{"module_size":[42.0,43.0]},"src/punt_vox/voxd/dedup.py":{"module_size":[165.0,175.0],"init_violations":[2.0,0.0]}}}
+{"ts":"2026-05-14T01:29:04Z","commit":"0b49063","files_scored":41,"files_improved":3,"files_regressed":1,"verdict":"fail","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[110.0,114.0]},"src/punt_vox/voxd/playback.py":{"method_ratio":[0.0,0.5],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.67],"max_complexity":[0.0,13.0],"avg_complexity":[0.0,2.22],"module_size":[0.0,346.0],"classes_per_module":[0.0,2.0],"class_to_func_ratio":[0.0,0.182],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/track_generator.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.33],"max_complexity":[0.0,5.0],"avg_complexity":[0.0,2.33],"module_size":[0.0,73.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T01:29:35Z","commit":"0b49063","files_scored":41,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.019,0.089],"encapsulation_ratio":[0.0,0.056],"avg_params":[2.23,2.43],"avg_complexity":[3.88,3.71],"module_size":[1922.0,1645.0],"classes_per_module":[3.0,2.0],"public_attr_violations":[19.0,17.0]}}}
+{"ts":"2026-05-14T01:31:26Z","commit":"0b49063","files_scored":41,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.019,0.089],"encapsulation_ratio":[0.0,0.056],"avg_params":[2.23,2.43],"avg_complexity":[3.88,3.71],"module_size":[1922.0,1645.0],"classes_per_module":[3.0,2.0],"public_attr_violations":[19.0,17.0]}}}
+{"ts":"2026-05-14T01:34:33Z","commit":"0b49063","files_scored":41,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.019,0.089]}}}
+{"ts":"2026-05-14T01:38:05Z","commit":"0b49063","files_scored":41,"files_improved":4,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[110.0,114.0]},"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.019,0.089],"encapsulation_ratio":[0.0,0.056],"avg_params":[2.23,2.43],"avg_complexity":[3.88,3.71],"module_size":[1922.0,1645.0],"classes_per_module":[3.0,2.0],"public_attr_violations":[19.0,17.0]},"src/punt_vox/voxd/playback.py":{"method_ratio":[0.0,0.5],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.67],"max_complexity":[0.0,13.0],"avg_complexity":[0.0,2.22],"module_size":[0.0,346.0],"classes_per_module":[0.0,2.0],"class_to_func_ratio":[0.0,0.182],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]},"src/punt_vox/voxd/track_generator.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.33],"max_complexity":[0.0,5.0],"avg_complexity":[0.0,2.33],"module_size":[0.0,73.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
