+++ .oo-baseline.json
@@ -434,7 +434,7 @@
     "avg_params": 0.0,
     "max_complexity": 0.0,
     "avg_complexity": 0.0,
-    "module_size": 110.0,
+    "module_size": 114.0,
     "classes_per_module": 0.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
@@ -442,16 +442,16 @@
     "future_annotations": 1.0
   },
   "src/punt_vox/voxd/_monolith.py": {
-    "method_ratio": 0.019,
-    "encapsulation_ratio": 0.0,
-    "avg_params": 2.23,
+    "method_ratio": 0.089,
+    "encapsulation_ratio": 0.056,
+    "avg_params": 2.43,
     "max_complexity": 20.0,
-    "avg_complexity": 3.88,
-    "module_size": 1922.0,
-    "classes_per_module": 3.0,
-    "class_to_func_ratio": 0.056,
+    "avg_complexity": 3.71,
+    "module_size": 1645.0,
+    "classes_per_module": 2.0,
+    "class_to_func_ratio": 0.047,
     "init_violations": 1.0,
-    "public_attr_violations": 19.0,
+    "public_attr_violations": 17.0,
     "future_annotations": 1.0
   },
   "src/punt_vox/voxd/chimes.py": {
@@ -488,6 +488,32 @@
     "avg_complexity": 3.2,
     "module_size": 175.0,
     "classes_per_module": 3.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/playback.py": {
+    "method_ratio": 0.5,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 0.67,
+    "max_complexity": 13.0,
+    "avg_complexity": 2.22,
+    "module_size": 346.0,
+    "classes_per_module": 2.0,
+    "class_to_func_ratio": 0.182,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/track_generator.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 1.33,
+    "max_complexity": 5.0,
+    "avg_complexity": 2.33,
+    "module_size": 73.0,
+    "classes_per_module": 1.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
     "public_attr_violations": 0.0,
