+++ pyproject.toml
@@ -111,6 +111,7 @@ ignore = [
 "src/punt_vox/__main__.py" = ["T20", "PLC0415", "S603", "S607"]
 "src/punt_vox/server.py" = ["PLC0415"]
 "src/punt_vox/voxd/_monolith.py" = ["PLC0415"]
+"src/punt_vox/voxd/track_generator.py" = ["PLC0415"]
 "src/punt_vox/applet.py" = ["PLC0415"]
 "src/punt_vox/client.py" = ["PLC0415"]
 "src/punt_vox/hooks.py" = ["T20", "PLC0415", "S603", "S607", "S311"]
