+++ src/punt_vox/voxd/track_generator.py
@@ -1,0 +1,95 @@
+"""Music track generation -- name derivation, slugification, and track listing."""
+
+from __future__ import annotations
+
+import logging
+import re
+import time
+from pathlib import Path
+from typing import Self
+
+__all__ = ["TrackGenerator"]
+
+logger = logging.getLogger(__name__)
+
+_MUSIC_DURATION_MS = 120_000
+
+
+class TrackGenerator:
+    """Generate, name, and list music tracks in an output directory."""
+
+    __slots__ = ("_output_dir",)
+
+    _output_dir: Path
+
+    def __new__(cls, output_dir: Path) -> Self:
+        self = super().__new__(cls)
+        self._output_dir = output_dir
+        return self
+
+    @property
+    def output_dir(self) -> Path:
+        """Return the output directory for generated tracks."""
+        return self._output_dir
+
+    async def generate(
+        self,
+        vibe: tuple[str, str],
+        style: str,
+        track_name: str,
+    ) -> tuple[Path, str]:
+        """Generate a music track and return (track_path, resolved_track_name).
+
+        Uses explicit vibe/style/track_name parameters rather than reading
+        from DaemonContext, keeping TrackGenerator decoupled from daemon state.
+        """
+        vibe_text, vibe_tags = vibe
+
+        from punt_vox.music import vibe_to_prompt
+
+        hour = time.localtime().tm_hour
+        prompt = vibe_to_prompt(
+            vibe_text or None, vibe_tags or None, style or None, hour, signals=[]
+        )
+
+        self._output_dir.mkdir(parents=True, exist_ok=True)
+
+        resolved_name = track_name or self.auto_track_name(vibe_text, style)
+        safe_name = self.slugify(resolved_name, max_len=60)
+        filename = f"{safe_name}.mp3"
+        output_path = self._output_dir / filename
+
+        from punt_vox.providers.elevenlabs_music import ElevenLabsMusicProvider
+
+        provider = ElevenLabsMusicProvider()
+        await provider.generate_track(prompt, _MUSIC_DURATION_MS, output_path)
+        return output_path, safe_name
+
+    def auto_track_name(self, vibe: str, style: str) -> str:
+        """Derive a short auto-name from vibe + style + YYYYMMDD-HHMM."""
+        stamp = time.strftime("%Y%m%d-%H%M")
+        vibe_part = self.slugify(vibe, max_len=20) or "ambient"
+        style_part = self.slugify(style, max_len=20) or "mix"
+        return f"{vibe_part}-{style_part}-{stamp}"
+
+    def list_tracks(self) -> list[dict[str, object]]:
+        """Return metadata for all saved .mp3 tracks in the output directory."""
+        tracks: list[dict[str, object]] = []
+        if self._output_dir.exists():
+            for mp3 in sorted(self._output_dir.glob("*.mp3")):
+                stat = mp3.stat()
+                tracks.append(
+                    {
+                        "name": mp3.stem,
+                        "size_bytes": stat.st_size,
+                        "modified": stat.st_mtime,
+                        "path": str(mp3),
+                    }
+                )
+        return tracks
+
+    @staticmethod
+    def slugify(text: str, max_len: int = 40) -> str:
+        """Slugify a string for use in filenames."""
+        slug = re.sub(r"[^a-zA-Z0-9]+", "_", text).strip("_").lower()
+        return slug[:max_len]
